+++ config/flatten.php
@@ -27,10 +27,6 @@ return [
     // Cache variables
     ////////////////////////////////////////////////////////////////////
 
-    // Whether to append a timestamp comment on cached pages
-    // eg. <!-- Cached on 2015-09-09 01:01:01 -->
-    'timestamp' => true,
-
     // The default period during which a cached page should be kept (in minutes)
     // 0 means the page never gets refreshed by itself
     'lifetime' => 0,
@@ -38,4 +34,15 @@ return [
     // An array of string or variables to add to the salt being used
     // to differentiate pages
     'saltshaker' => [],
+
+    // Cache output
+    //////////////////////////////////////////////////////////////////////
+
+    // Whether to append a timestamp comment on cached pages
+    // eg. <!-- Cached on 2015-09-09 01:01:01 -->
+    'timestamp' => true,
+
+    // Whether to minify the HTML before caching it
+    'minify' => false,
+
 ];